backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pyheif==0.7.1
pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
starlette==0.27.0
pytz==2023.3
requests==2.31.0
